make test_llm_api_pytorch_multimodal and move multimodal accuracy tests
