+++ tests/integration/defs/accuracy/test_llm_api_pytorch.py
@@ -32,7 +32,7 @@ from ..conftest import (get_device_count, get_device_memory, llm_models_root,
                         parametrize_with_ids, skip_no_hopper,
                         skip_post_blackwell, skip_pre_ada, skip_pre_blackwell,
                         skip_pre_hopper, skip_ray)
-from .accuracy_core import (GSM8K, MMLU, MMMU, CnnDailymail, GPQADiamond,
+from .accuracy_core import (GSM8K, MMLU, CnnDailymail, GPQADiamond,
                             JsonModeEval, LlmapiAccuracyTestHarness)
 
 
@@ -3880,46 +3880,6 @@ class TestEXAONE4(LlmapiAccuracyTestHarness):
             task.evaluate(llm)
 
 
-class TestQwen2_VL_7B(LlmapiAccuracyTestHarness):
-    MODEL_NAME = "Qwen/Qwen2-VL-7B-Instruct"
-    MODEL_PATH = f"{llm_models_root()}/Qwen2-VL-7B-Instruct"
-    MAX_NUM_TOKENS = 16384
-    # NOTE: MMMU adds <|endoftext|> to the stop token.
-    sampling_params = SamplingParams(max_tokens=MMMU.MAX_OUTPUT_LEN,
-                                     truncate_prompt_tokens=MMMU.MAX_INPUT_LEN,
-                                     stop="<|endoftext|>")
-
-    kv_cache_config = KvCacheConfig(free_gpu_memory_fraction=0.6)
-
-    @pytest.mark.skip(reason="https://nvbugs/5601909")
-    def test_auto_dtype(self):
-        with LLM(self.MODEL_PATH,
-                 max_num_tokens=self.MAX_NUM_TOKENS,
-                 kv_cache_config=self.kv_cache_config) as llm:
-            task = MMMU(self.MODEL_NAME)
-            task.evaluate(llm, sampling_params=self.sampling_params)
-
-
-class TestQwen2_5_VL_7B(LlmapiAccuracyTestHarness):
-    MODEL_NAME = "Qwen/Qwen2.5-VL-7B-Instruct"
-    MODEL_PATH = f"{llm_models_root()}/Qwen2.5-VL-7B-Instruct"
-    MAX_NUM_TOKENS = 16384
-
-    # NOTE: MMMU adds <|endoftext|> to the stop token.
-    sampling_params = SamplingParams(max_tokens=MMMU.MAX_OUTPUT_LEN,
-                                     truncate_prompt_tokens=MMMU.MAX_INPUT_LEN,
-                                     stop="<|endoftext|>")
-
-    kv_cache_config = KvCacheConfig(free_gpu_memory_fraction=0.6)
-
-    def test_auto_dtype(self):
-        with LLM(self.MODEL_PATH,
-                 max_num_tokens=self.MAX_NUM_TOKENS,
-                 kv_cache_config=self.kv_cache_config) as llm:
-            task = MMMU(self.MODEL_NAME)
-            task.evaluate(llm, sampling_params=self.sampling_params)
-
-
 class TestQwQ_32B(LlmapiAccuracyTestHarness):
     MODEL_NAME = "Qwen/QwQ-32B"
     MODEL_PATH = f"{llm_models_root()}/QwQ-32B"
@@ -3971,29 +3931,6 @@ class TestQwen3NextThinking(LlmapiAccuracyTestHarness):
             task.evaluate(llm)
 
 
-class TestNano_V2_VLM(LlmapiAccuracyTestHarness):
-    MODEL_NAME = "nvidia/Nano-v2-VLM"
-    MODEL_PATH = f"{llm_models_root()}/Nano-v2-VLM"
-    MAX_NUM_TOKENS = 25600
-
-    # NOTE: MMMU adds <|endoftext|> to the stop token.
-    sampling_params = SamplingParams(max_tokens=MAX_NUM_TOKENS,
-                                     truncate_prompt_tokens=MMMU.MAX_INPUT_LEN,
-                                     stop="<|endoftext|>")
-
-    kv_cache_config = KvCacheConfig(free_gpu_memory_fraction=0.8,
-                                    enable_block_reuse=False)
-
-    @pytest.mark.skip(reason="Nano V2 VLM ckpt is not released yet.")
-    def test_auto_dtype(self):
-        with LLM(self.MODEL_PATH,
-                 max_batch_size=128,
-                 max_num_tokens=self.MAX_NUM_TOKENS,
-                 kv_cache_config=self.kv_cache_config) as llm:
-            task = MMMU(self.MODEL_NAME)
-            task.evaluate(llm, sampling_params=self.sampling_params)
-
-
 class TestSeedOss_36B(LlmapiAccuracyTestHarness):
     MODEL_NAME = "ByteDance-Seed/Seed-OSS-36B-Instruct"
     MODEL_PATH = f"{llm_models_root()}/Seed-OSS/Seed-OSS-36B-Instruct"
@@ -4015,23 +3952,3 @@ class TestSeedOss_36B(LlmapiAccuracyTestHarness):
                           extra_evaluator_kwargs=dict(
                               apply_chat_template=True,
                               chat_template_kwargs=chat_template_kwargs))
-
-
-class TestLlava_V1_6_Mistral_7B(LlmapiAccuracyTestHarness):
-    MODEL_NAME = "llava-hf/llava-v1.6-mistral-7b"
-    MODEL_PATH = f"{llm_models_root()}/llava-v1.6-mistral-7b"
-    MAX_NUM_TOKENS = 16384
-
-    # NOTE: MMMU adds <|endoftext|> to the stop token.
-    sampling_params = SamplingParams(max_tokens=MMMU.MAX_OUTPUT_LEN,
-                                     truncate_prompt_tokens=MMMU.MAX_INPUT_LEN,
-                                     stop="<|endoftext|>")
-
-    kv_cache_config = KvCacheConfig(free_gpu_memory_fraction=0.6)
-
-    def test_auto_dtype(self):
-        with LLM(self.MODEL_PATH,
-                 max_num_tokens=self.MAX_NUM_TOKENS,
-                 kv_cache_config=self.kv_cache_config) as llm:
-            task = MMMU(self.MODEL_NAME)
-            task.evaluate(llm, sampling_params=self.sampling_params)
