+++ tests/integration/test_lists/qa/llm_function_gb20x.txt
@@ -21,6 +21,7 @@ examples/test_enc_dec.py::test_llm_enc_dec_general[compare_hf-flan-t5-small-floa
 examples/test_enc_dec.py::test_llm_enc_dec_general[compare_hf-t5-small-float32-enable_gemm_plugin-enable_attention_plugin-enable_paged_kv_cache-tp:1-pp:1-nb:1-disable_fp8]
 accuracy/test_cli_flow.py::TestStarcoder2_3B::test_auto_dtype
 accuracy/test_cli_flow.py::TestStarcoder2_15B::test_smooth_quant_ootb
-accuracy/test_llm_api_pytorch.py::TestQwen2_VL_7B::test_auto_dtype
-accuracy/test_llm_api_pytorch.py::TestQwen2_5_VL_7B::test_auto_dtype
-accuracy/test_llm_api_pytorch.py::TestLlava_V1_6_Mistral_7B::test_auto_dtype
+
+accuracy/test_llm_api_pytorch_multimodal.py::TestQwen2_VL_7B::test_auto_dtype
+accuracy/test_llm_api_pytorch_multimodal.py::TestQwen2_5_VL_7B::test_auto_dtype
+accuracy/test_llm_api_pytorch_multimodal.py::TestLlava_V1_6_Mistral_7B::test_auto_dtype
