+++ tests/integration/test_lists/qa/llm_function_l20.txt
@@ -45,9 +45,10 @@ accuracy/test_llm_api_pytorch.py::TestPhi4MM::test_fp4
 accuracy/test_llm_api_pytorch.py::TestPhi4MM::test_fp8
 accuracy/test_llm_api_pytorch.py::TestPhi4MiniInstruct::test_auto_dtype
 accuracy/test_llm_api_pytorch.py::TestMistralNemo12B::test_auto_dtype
-accuracy/test_llm_api_pytorch.py::TestQwen2_VL_7B::test_auto_dtype
-accuracy/test_llm_api_pytorch.py::TestQwen2_5_VL_7B::test_auto_dtype
-accuracy/test_llm_api_pytorch.py::TestLlava_V1_6_Mistral_7B::test_auto_dtype
+
+accuracy/test_llm_api_pytorch_multimodal.py::TestQwen2_VL_7B::test_auto_dtype
+accuracy/test_llm_api_pytorch_multimodal.py::TestQwen2_5_VL_7B::test_auto_dtype
+accuracy/test_llm_api_pytorch_multimodal.py::TestLlava_V1_6_Mistral_7B::test_auto_dtype
 
 
 test_e2e.py::test_ptp_quickstart_multimodal[NVILA-8B-FP16-vila/NVILA-8B-image-False]
